+++ lib/src/native/realm_core.dart
@@ -54,7 +54,7 @@ const bugInTheSdkMessage = "This is likely a bug in the Realm SDK - please file 
 
 final _realmLib = () {
   final result = RealmLibrary(initRealm());
-  nativeLibraryVersion = result.realm_dart_library_version().cast<Utf8>().toDartString();
+  final nativeLibraryVersion = result.realm_dart_library_version().cast<Utf8>().toDartString();
   if (libraryVersion != nativeLibraryVersion) {
     final additionMessage =
         isFlutterPlatform ? bugInTheSdkMessage : "Did you forget to run `dart run realm_dart install` after upgrading the realm_dart package?";
@@ -65,8 +65,6 @@ final _realmLib = () {
 
 // stamped into the library by the build system (see prepare-release.yml)
 const libraryVersion = '1.0.3';
-late String nativeLibraryVersion;
-
 final realmCore = _RealmCore();
 
 class _RealmCore {
