+++ src/CMakeLists.txt
@@ -9,6 +9,7 @@ add_subdirectory(dart-dl)
 
 set(SOURCES
     realm_dart.cpp
+    realm_dart_decimal128.cpp
     realm_dart_scheduler.cpp
     realm_dart_sync.cpp
     realm-core/src/external/IntelRDFPMathLib20U2/LIBRARY/src/bid128_noncomp.c
@@ -17,6 +18,7 @@ set(SOURCES
 set(HEADERS
     realm_dart.h
     realm_dart.hpp
+    realm_dart_scheduler.h
     realm_dart_scheduler.h
     realm_dart_sync.h
     realm-core/src/realm.h
